pyproject.toml
[tool.poetry]
name = "drf-online-learning"
version = "0.1.0"
description = ""
authors = ["Yulia Samoylova <u.u.samoylova@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.2.1"
djangorestframework = "^3.16.0"
black = "^25.1.0"
flake8 = "^7.2.0"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.1.0"
pillow = "^11.2.1"
django-filter = "^25.1"
djangorestframework-simplejwt = "^5.5.0"
ipython = "^9.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
